+++ src/IncrementalInference.jl
@@ -14,7 +14,7 @@ using Reexport
 
 using Manifolds
 
-export ℝ, Manifold, Euclidean, Circle
+export ℝ, AbsstractManifold, Euclidean, Circle
 
 import NLsolve
 import NLSolversBase
@@ -74,6 +74,7 @@ import DistributedFactorGraphs: isSolvable
 
 # must be moved to their own repos
 const KDE = KernelDensityEstimate
+const MB = ManifoldsBase
 const AMP = ApproxManifoldProducts
 const FSM = FunctionalStateMachine
 const IIF = IncrementalInference